# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CodeMaster coding-practice platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create a CodeMaster account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/ApiError"
        "409":
          $ref: "#/components/responses/ApiError"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to CodeMaster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          $ref: "#/components/responses/ApiError"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the signed-in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/ApiError"
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: Search and filter problems
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema: { type: string, enum: [Easy, Medium, Hard] }
        - name: topic
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        "200":
          description: Filtered problem page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemPage"
  /problems/{slug}:
    get:
      operationId: getProblem
      tags: [problems]
      summary: Get a problem and its public examples
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Problem details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          $ref: "#/components/responses/ApiError"
  /submissions/run:
    post:
      operationId: runCode
      tags: [submissions]
      summary: Run code against a custom input in an isolated runner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeRunInput"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"
        "400":
          $ref: "#/components/responses/ApiError"
        "503":
          $ref: "#/components/responses/ApiError"
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List the signed-in user's submissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [Accepted, Wrong Answer, Time Limit Exceeded, Runtime Error, Compilation Error, Internal Error]
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
      responses:
        "200":
          description: Submission history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionPage"
        "401":
          $ref: "#/components/responses/ApiError"
    post:
      operationId: submitCode
      tags: [submissions]
      summary: Submit code for evaluation against public and hidden tests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Evaluated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "401":
          $ref: "#/components/responses/ApiError"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission owned by the current user
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          $ref: "#/components/responses/ApiError"
...
components:
  responses:
    ApiError:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        avatar: { type: ["string", "null"] }
        role: { type: string, enum: [user, admin] }
        solvedCount: { type: integer }
        currentStreak: { type: integer }
        totalSubmissions: { type: integer }
        acceptedSubmissions: { type: integer }
        joinedAt: { type: string, format: date-time }
      required: [id, username, name, email, avatar, role, solvedCount, currentStreak, totalSubmissions, acceptedSubmissions, joinedAt]
    AuthResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        message: { type: string }
      required: [user, message]
    RegistrationInput:
      type: object
      properties:
        username: { type: string, minLength: 3, maxLength: 24, pattern: "^[a-zA-Z0-9_]+$" }
        name: { type: string, minLength: 1, maxLength: 80 }
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 128 }
      required: [username, name, email, password]
    CredentialsInput:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
      required: [email, password]
    Example:
      type: object
      properties:
        input: { type: string }
        output: { type: string }
        explanation: { type: ["string", "null"] }
      required: [input, output, explanation]
...
    ProblemSummary:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        difficulty: { type: string, enum: [Easy, Medium, Hard] }
        topics:
          type: array
          items: { type: string }
        acceptanceRate: { type: number }
        totalSubmissions: { type: integer }
        acceptedSubmissions: { type: integer }
        solved: { type: boolean }
      required: [id, slug, title, difficulty, topics, acceptanceRate, totalSubmissions, acceptedSubmissions, solved]
    Problem:
      type: object
      properties:
        id: { type: integer }
        slug: { type: string }
        title: { type: string }
        description: { type: string }
        difficulty: { type: string, enum: [Easy, Medium, Hard] }
        topics:
          type: array
          items: { type: string }
        constraints:
          type: array
          items: { type: string }
        examples:
          type: array
          items: { $ref: "#/components/schemas/Example" }
        starterCode:
          type: object
          additionalProperties: { type: string }
        supportedLanguages:
          type: array
          items: { type: string }
        acceptanceRate: { type: number }
        totalSubmissions: { type: integer }
        acceptedSubmissions: { type: integer }
        solved: { type: boolean }
      required: [id, slug, title, description, difficulty, topics, constraints, examples, starterCode, supportedLanguages, acceptanceRate, totalSubmissions, acceptedSubmissions, solved]
    ProblemPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ProblemSummary" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }
      required: [items, page, limit, total, pages]
    CodeRunInput:
      type: object
      properties:
        problemId: { type: integer, minimum: 1 }
        language: { type: string, enum: [javascript, python, java, cpp] }
        code: { type: string, minLength: 1, maxLength: 50000 }
        input: { type: string, maxLength: 10000 }
      required: [problemId, language, code, input]
    RunResult:
      type: object
      properties:
        status: { type: string }
        output: { type: ["string", "null"] }
        error: { type: ["string", "null"] }
        runtime: { type: ["number", "null"] }
        memory: { type: ["number", "null"] }
      required: [status, output, error, runtime, memory]
    SubmissionInput:
      type: object
      properties:
        problemId: { type: integer, minimum: 1 }
        language: { type: string, enum: [javascript, python, java, cpp] }
        code: { type: string, minLength: 1, maxLength: 50000 }
      required: [problemId, language, code]
    Submission:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        problemId: { type: integer }
        problemTitle: { type: string }
        problemSlug: { type: string }
        language: { type: string }
        code: { type: string }
        status: { type: string, enum: [Accepted, Wrong Answer, "Time Limit Exceeded", "Runtime Error", "Compilation Error", "Internal Error"] }
        runtime: { type: ["number", "null"] }
        memory: { type: ["number", "null"] }
        testCasesPassed: { type: integer }
        totalTestCases: { type: integer }
        errorMessage: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
      required: [id, userId, problemId, problemTitle, problemSlug, language, code, status, runtime, memory, testCasesPassed, totalTestCases, errorMessage, submittedAt]
...
    SubmissionPage:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Submission" }
        page: { type: integer }
        total: { type: integer }
        pages: { type: integer }
      required: [items, page, total, pages]
...