# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Example } from './example';
import type { ProblemDifficulty } from './problemDifficulty';
import type { ProblemStarterCode } from './problemStarterCode';

export interface Problem {
  id: number;
  slug: string;
  title: string;
  description: string;
  difficulty: ProblemDifficulty;
  topics: string[];
  constraints: string[];
  examples: Example[];
  starterCode: ProblemStarterCode;
  supportedLanguages: string[];
  acceptanceRate: number;
  totalSubmissions: number;
  acceptedSubmissions: number;
  solved: boolean;
}