# lib/api-zod/src/generated/types/problemPage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemSummary } from './problemSummary';

export interface ProblemPage {
  items: ProblemSummary[];
  page: number;
  limit: number;
  total: number;
  pages: number;
}