# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /**
     * @minLength 3
     * @maxLength 24
     * @pattern ^[a-zA-Z0-9_]+$
     */
  username: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}