# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * CodeMaster coding-practice platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  username: string;
  name: string;
  email: string;
  /** @nullable */
  avatar: string | null;
  role: UserRole;
  solvedCount: number;
  currentStreak: number;
  totalSubmissions: number;
  acceptedSubmissions: number;
  joinedAt: Date;
}